refactor search tests
