--- tests/search_where_edge_test.rs
@@ -1,10 +1,0 @@
-use agdb::QueryBuilder;
-
-#[test]
-fn search_from_where_edge() {
-    let _query = QueryBuilder::search()
-        .from(1.into())
-        .where_()
-        .edge()
-        .query();
-}
