--- tests/search_where_ids_test.rs
@@ -1,19 +1,0 @@
-use agdb::QueryBuilder;
-
-#[test]
-fn search_from_where_id() {
-    let _query = QueryBuilder::search()
-        .from(1.into())
-        .where_()
-        .id("alias".into())
-        .query();
-}
-
-#[test]
-fn search_from_where_ids() {
-    let _query = QueryBuilder::search()
-        .from(1.into())
-        .where_()
-        .ids(&["alias".into(), "alias2".into()])
-        .query();
-}
