+++ tests/select_aliases_test.rs
@@ -1,11 +1,6 @@
 use agdb::QueryBuilder;
 
 #[test]
-fn select_alias_id() {
-    let _query = QueryBuilder::select().aliases().ids(&[1.into()]).query();
-}
-
-#[test]
 fn select_aliases() {
     let _query = QueryBuilder::select().aliases().query();
 }
