+++ tests/select_key_count_test.rs
@@ -1,14 +1,6 @@
 use agdb::QueryBuilder;
 
 #[test]
-fn select_key_count_id() {
-    let _query = QueryBuilder::select()
-        .key_count()
-        .id("alias".into())
-        .query();
-}
-
-#[test]
 fn select_key_count_ids() {
     let _query = QueryBuilder::select()
         .key_count()
