+++ tests/select_keys.rs
@@ -1,11 +1,6 @@
 use agdb::QueryBuilder;
 
 #[test]
-fn select_keys_id() {
-    let _query = QueryBuilder::select().keys().id("alias".into()).query();
-}
-
-#[test]
 fn select_keys_ids() {
     let _query = QueryBuilder::select().keys().ids(&["alias".into()]).query();
 }
